combine test workers
