--- crates/test-framework/src/bench/mod.rs
@@ -1,170 +1,0 @@
-/*
-Copyright 2024-2025 The Spice.ai OSS Authors
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-     https://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-use std::{
-    collections::BTreeMap,
-    time::{Duration, Instant, SystemTime},
-};
-
-use crate::spiced::SpicedInstance;
-use anyhow::Result;
-use tokio::task::JoinHandle;
-use worker::{BenchQueryWorker, BenchQueryWorkerResult};
-
-mod worker;
-
-pub struct NotStarted {
-    query_set: Vec<(&'static str, &'static str)>,
-}
-
-pub type QueryWorker = JoinHandle<Result<BenchQueryWorkerResult>>;
-
-pub struct Running {
-    start_time: Instant,
-    query_worker: QueryWorker,
-}
-pub struct Completed {
-    query_durations: BTreeMap<String, Vec<Duration>>,
-    test_duration: Duration,
-    end_time: SystemTime,
-}
-
-pub trait TestState {}
-
-impl TestState for NotStarted {}
-impl TestState for Running {}
-impl TestState for Completed {}
-
-pub struct BenchTest<S: TestState> {
-    name: String,
-    spiced_instance: SpicedInstance,
-    query_count: usize,
-    start_time: SystemTime,
-    iterations: usize,
-
-    state: S,
-}
-
-impl BenchTest<NotStarted> {
-    #[must_use]
-    pub fn new(name: String, spiced_instance: SpicedInstance) -> Self {
-        Self {
-            name,
-            spiced_instance,
-            query_count: 0,
-            start_time: SystemTime::now(),
-            iterations: 5,
-            state: NotStarted { query_set: vec![] },
-        }
-    }
-
-    #[must_use]
-    pub fn with_iterations(mut self, iterations: usize) -> Self {
-        self.iterations = iterations;
-        self
-    }
-
-    #[must_use]
-    pub fn with_query_set(mut self, query_set: Vec<(&'static str, &'static str)>) -> Self {
-        self.query_count = query_set.len();
-        self.state.query_set = query_set;
-        self
-    }
-
-    pub async fn start(self) -> Result<BenchTest<Running>> {
-        if self.state.query_set.is_empty() {
-            return Err(anyhow::anyhow!("Query set is empty"));
-        }
-
-        if self.iterations == 0 {
-            return Err(anyhow::anyhow!("Iterations count must be greater than 0"));
-        }
-
-        let flight_client = self.spiced_instance.flight_client().await?;
-        let query_worker = BenchQueryWorker::new(
-            self.iterations,
-            flight_client.clone(),
-            self.state.query_set.clone(),
-        )
-        .start();
-
-        Ok(BenchTest {
-            name: self.name,
-            spiced_instance: self.spiced_instance,
-            query_count: self.query_count,
-            iterations: self.iterations,
-            start_time: self.start_time,
-            state: Running {
-                start_time: Instant::now(),
-                query_worker,
-            },
-        })
-    }
-}
-
-impl BenchTest<Running> {
-    pub async fn wait(self) -> Result<BenchTest<Completed>> {
-        let query_duration = self.state.query_worker.await;
-        let worker_result = query_duration??;
-        if worker_result.connection_failed {
-            return Err(anyhow::anyhow!(
-                "Test failed - a connection failed during the test"
-            ));
-        }
-
-        Ok(BenchTest {
-            name: self.name,
-            spiced_instance: self.spiced_instance,
-            query_count: self.query_count,
-            iterations: self.iterations,
-            start_time: self.start_time,
-            state: Completed {
-                query_durations: worker_result.query_durations,
-                test_duration: self.state.start_time.elapsed(),
-                end_time: SystemTime::now(),
-            },
-        })
-    }
-}
-
-impl BenchTest<Completed> {
-    #[must_use]
-    pub fn get_query_durations(&self) -> &BTreeMap<String, Vec<Duration>> {
-        &self.state.query_durations
-    }
-
-    #[must_use]
-    pub fn get_test_duration(&self) -> Duration {
-        self.state.test_duration
-    }
-
-    /// Once the test has completed, return ownership of the spiced instance
-    #[must_use]
-    pub fn end(self) -> SpicedInstance {
-        self.spiced_instance
-    }
-}
-
-impl std::fmt::Display for BenchTest<Completed> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            "BenchTest: {} - Test duration: {} seconds",
-            self.name,
-            self.get_test_duration().as_secs_f32()
-        )
-    }
-}
